Remove temporary tracing middleware
