+++ tower-http/src/classify.rs
@@ -219,3 +219,63 @@ fn classify_grpc_metadata(headers: &HeaderMap) -> Option<Result<(), i32>> {
         Some(Err(status))
     }
 }
+
+// Just verify that we can actually use this response classifier to determine retries as well
+#[cfg(test)]
+mod usable_for_retries {
+    #[allow(unused_imports)]
+    use super::*;
+    use hyper::{Request, Response};
+    use tower::retry::Policy;
+
+    trait IsRetryable {
+        fn is_retryable(&self) -> bool;
+    }
+
+    #[derive(Clone)]
+    struct RetryBasedOnClassification<C> {
+        classifier: C,
+        // ...
+    }
+
+    impl<ReqB, ResB, E, C> Policy<Request<ReqB>, Response<ResB>, E> for RetryBasedOnClassification<C>
+    where
+        C: ClassifyResponse<E> + Clone,
+        C::FailureClass: IsRetryable,
+        ResB: http_body::Body,
+        Request<ReqB>: Clone,
+        E: std::error::Error + 'static,
+    {
+        type Future = futures::future::Ready<RetryBasedOnClassification<C>>;
+
+        fn retry(
+            &self,
+            _req: &Request<ReqB>,
+            res: Result<&Response<ResB>, &E>,
+        ) -> Option<Self::Future> {
+            match res {
+                Ok(res) => {
+                    if let ClassifiedResponse::Ready(class) =
+                        self.classifier.clone().classify_response(res)
+                    {
+                        if class.err()?.is_retryable() {
+                            return Some(futures::future::ready(self.clone()));
+                        }
+                    }
+
+                    None
+                }
+                Err(err) => self
+                    .classifier
+                    .clone()
+                    .classify_error(err)
+                    .is_retryable()
+                    .then(|| futures::future::ready(self.clone())),
+            }
+        }
+
+        fn clone_request(&self, req: &Request<ReqB>) -> Option<Request<ReqB>> {
+            Some(req.clone())
+        }
+    }
+}
